tema batch:  Frontend for model configuration and test execution.
tema batch is a small program for model configuration and test
execution. It reads its input from the stdin and the directory
'Model'.  Output is written into the directory 'TestModel' and the
file 'engine.prms' in the working directory.  Optionally it can start
test generation right after the model is configured.  Execute 'tema
batch --help' for documentation.